ai/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
google-generativeai==0.3.0
python-multipart==0.0.6
bcrypt==4.0.1
python-jose==3.3.0
passlib==1.7.4
